backend/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
flask-socketio>=5.0.0
python-dotenv>=0.19.0
requests>=2.25.0
openai>=0.27.0
gunicorn>=20.0.0
eventlet>=0.30.0 